Fixed comment controller bugs

+++ src/controllers/comment.controller.js
@@ -8,76 +8,86 @@ const getVideoComments = asyncHandler(async (req, res) => {
   //TODO: get all comments for a video
   const { videoId } = req.params;
   const { page = 1, limit = 10 } = req.query;
-
+  
   if (!videoId) {
     throw new ApiError(400, "Invalid video id");
   }
-  const commentAggregate = await Comment.aggregate([
-    {
-      $match: {
-        video: new mongoose.Types.ObjectId(videoId),
-      },
-    },
-    {
-      $lookup: {
-        from: "users",
-        localField: "owner",
-        foreignField: "_id",
-        as: "owner",
-        pipeline: [
-          {
-            $project: {
-              _id: 1,
-              fullname: 1,
-              username: 1,
-              avatar: 1,
-            },
-          },
-        ],
-      },
-    },
-    {
-      $addFields: {
-        owner: {
-          $first: "$owner",
+  try {
+    const commentAggregate = await Comment.aggregate([
+      {
+        $match: {
+          video: new mongoose.Types.ObjectId(videoId),
         },
       },
-    },
-  ]);
-
-  const options = {
-    page,
-    limit,
-    customLabels: {
-      totalDocs: "totalVideos",
-      docs: "videos",
-    },
-    skip: (page - 1) * limit,
-    limit: parseInt(limit),
-  };
-
-  const comments = await Comment.aggregatePaginate(commentAggregate, options);
-  if (!comments) {
-    throw new ApiError(400, "An error occured while fetching the comments");
+      {
+        $lookup: {
+          from: "users",
+          localField: "owner",
+          foreignField: "_id",
+          as: "owner",
+          pipeline: [
+            {
+              $project: {
+                _id: 1,
+                fullname: 1,
+                username: 1,
+                avatar: 1,
+              },
+            },
+          ],
+        },
+      },
+      {
+        $addFields: {
+          owner: {
+            $first: "$owner",
+          },
+        },
+      },
+      
+    ]);
+  
+    const page=1
+    const limit=20
+    const options = {
+      page:parseInt(page),
+      limit:parseInt(limit),
+      customLabels: {
+        totalDocs: "totalComments",
+        docs: "comments",
+      },
+      skip: (page - 1) * limit,
+      // limit: parseInt(limit),
+    };
+  
+    const comments = await Comment.aggregatePaginate(commentAggregate, options);
+    if (!comments) {
+      throw new ApiError(400, "An error occured while fetching the comments");
+    }
+  
+    return res
+      .status(201)
+      .json(
+        new ApiResponse(
+          200,
+          commentAggregate,
+          "Comments for this video fetched successfully"
+        )
+      );
+  } catch (error) {
+      console.log(error.message)
+      throw new ApiError(404,"Failed to paginate comments")
   }
-
-  return res
-    .status(201)
-    .json(
-      new ApiResponse(
-        200,
-        comments,
-        "Comments for this video fetched successfully"
-      )
-    );
 });
 
 const addComment = asyncHandler(async (req, res) => {
     const {videoId}=req.params
+    console.log(req.body)
     const {content}=req.body
-    if(!user.req?._id){
+    if(!req.user?._id){
         throw new ApiError(400,"You must be logged in to post a comment")
     }
+    console.log(content)
     if(!content){
         throw new ApiError(400,"Error while getting the content from body")
     }
@@ -103,7 +113,7 @@ const addComment = asyncHandler(async (req, res) => {
 
 const updateComment = asyncHandler(async (req, res) => {
     const {commentId}=req.params
-    const {newcontent}=req.body
+    const {content:newcontent}=req.body
 
     if(!commentId){
         throw new ApiError(400,"Error while fetching the comment to be updated")
@@ -133,7 +143,14 @@ const deleteComment = asyncHandler(async (req, res) => {
     if(!commentId){
         throw new ApiError(400,"Invalid delete comment Id")
     }
-
+    // console.log(commentId);
+    // const cid=new mongoose.Types.ObjectId(commentId)
+    // console.log(req.user._id);
+    const comment=await Comment.findOne({_id:commentId,owner:req.user?._id})
+    // console.log(comment);
+    if(!comment){
+      throw new ApiError(400,"You can only delete your own comments")
+    }
     const deletedComment=await Comment.findByIdAndDelete(commentId,{new:true})
 
     if(!deletedComment){
